batch_interface: move batch_context_verify and batch_context_add_schnorrsig
batch_context_verify:
- moved this API from module/schnorrsig/main_impl.h to secp256k1.c
- hence this API will be available even if the user doesn't enable schnorrsig module

batch_context_add_schnorrsig:
- moved this API from module/schnorrsig/main_impl.h to module/schnorrsig/batch_add_impl.h
- required since this api needs the `schnorrsig_challenge` func defined in schnorrsig/main_impl.h
- hence batch_impl.h will be included after schnorrsig/main_impl.h in the secp256k1.c file
